pyproject.toml
[tool.poetry]
name = "sellout"
version = "0.1.0"
description = ""
authors = ["Alexander Schepetilnikov <shchepetilnikov@tochka.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
pydantic = "^2.10.6"
gunicorn = "^23.0.0"
psycopg-binary = "^3.2.4"
psycopg2-binary = "^2.9.10"
drf-yasg = "^1.21.8"


[tool.poetry.group.dev.dependencies]
pre-commit = "4.1.0"
autoflake = "2.3.1"
black = "24.4.2"
flake8 = "7.0.0"
flake8-plugin-utils = "1.3.3"
flake8-pytest-style = "2.0.0"
flake8-tidy-imports = "4.10.0"
freezegun = "1.5.1"
pytest = "8.3.3"
mypy-extensions = "1.0.0"
pathspec = "0.12.1"
platformdirs = "4.2.2"
mypy = "1.10.0"
pytest-cov = "5.0.0"
pytest-asyncio = "0.24.0"
filelock = "3.14.0"
distlib = "0.3.8"
pytest-xdist = "3.6.1"
click = "8.1.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
exclude = ["venv", ".venv", "tests"]
strict = false
plugins = ["pydantic.mypy"]
disallow_untyped_defs = false
ignore_missing_imports = true
show_error_codes = true
show_column_numbers = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = "."
norecursedirs = [".*", "__pycache__", "venv", "static", ".venv"]

[tool.coverage.run]
omit = ["*/versions/*", "*/__main__.py", "*/tests/*", "*/fixtures/*"]